chore: consolidate FeelFunction logic into a module
Move all FEEL <function> concerns into a dedicated src/function.ts,
mirroring src/range.ts: the FeelFunction wrapper (data + invoke), the
isFunction guard, the FUNCTION_PARAMETER_MISMATCH sentinel, JS parameter
parsing (parseParameterNames) and the wrapFunction coercion factory
(the createRange analog) now live in one place. The interpreter keeps
only the function AST node handling and calls wrapFunction.
