+++ src/utils.ts
@@ -3,29 +3,6 @@ import { normalizeContextKey } from 'lezer-feel';
 import { getType } from './types.js';
 
 
-export function parseParameterNames(fn) {
-
-  if (Array.isArray(fn.$args)) {
-    return fn.$args;
-  }
-
-  const code = fn.toString();
-
-  const match = /^(?:[^(]*\s*)?\(([^)]+)?\)/.exec(code);
-
-  if (!match) {
-    throw new Error('failed to parse params: ' + code);
-  }
-
-  const [ _, params ] = match;
-
-  if (!params) {
-    return [];
-  }
-
-  return params.split(',').map(p => p.trim());
-}
-
 export function notImplemented(thing) {
   return new Error(`not implemented: ${thing}`);
 }
